+++ src/actions/runAllChangesets/getPendingChangesets.js
@@ -2,15 +2,28 @@ import fs from "fs";
 import path from "path";
 import extractDateFromString from "../../utils/extractDateFromString";
 import chalk from "chalk";
+import { isSomeArray } from "@locustjs/base";
 
-function getPendingChangesets(config, lastExecutedChangeset, executedChangesets) {
+function getPendingChangesets(config, executedChangesets) {
     console.log("Getting pending changesets ...");
+
+    const firstExecutedChangeset = isSomeArray(executedChangesets) ? executedChangesets[0] : null;
+    const lastExecutedChangeset = isSomeArray(executedChangesets) ? executedChangesets[executedChangesets.length - 1] : null;
+
+    if (firstExecutedChangeset) {
+        console.log(`First ${lastExecutedChangeset == firstExecutedChangeset ? "and Last " : ""}executed changeset: ${chalk.cyan(firstExecutedChangeset.name)}.`)
+
+        if (lastExecutedChangeset && lastExecutedChangeset != firstExecutedChangeset) {
+            console.log(`Last executed changeset: ${chalk.cyan(lastExecutedChangeset.name)}.`)
+        }
+    }
 
     const result = [];
     const files = fs.readdirSync(config.paths.changesetsPath);
     const changesets = files
         .filter(filepath => path.extname(filepath) == ".txt" && extractDateFromString(config, filepath))
         .map(filepath => path.parse(filepath).name)
+        .filter(filename => filename.match(/^(\d{14})/))
         .map(name => ({
             name,
             path: path.join(config.paths.changesetsPath, name + ".txt"),
@@ -18,49 +31,96 @@ function getPendingChangesets(config, lastExecutedChangeset, executedChangesets)
             date: extractDateFromString(config, name)
         }));
 
-    // ensure changesets that are older than lastExecutedChangeset will be also executed on database.
-    // this happens when we ahve two or more teams who have distinct workflows (each team has their
-    // own dev branch on which they merge their branches with).
-    console.log("\tchecking older changesets ...");
+    changesets.sort((a, b) => a.date - b.date);
 
-    changesets.forEach(changeset => {
-        if (!executedChangesets.find(cs => changeset.name.equals(cs.name))) {
-            config.debug2(`\t\tadded changeset ${chalk.yellow(changeset.name)}`);
-
-            result.push(changeset);
-        }
-    });
+    const firstExecutedChangesetName = firstExecutedChangeset?.name;
+    const firstExecutedDate = firstExecutedChangesetName ? extractDateFromString(config, firstExecutedChangesetName) : null;
 
     const lastExecutedChangesetName = lastExecutedChangeset?.name;
     const lastExecutedDate = lastExecutedChangesetName ? extractDateFromString(config, lastExecutedChangesetName) : null;
 
-    console.log("\tchecking newer changesets ...");
+    if (!firstExecutedChangeset) {
+        // database journal is empty.
+        // we should execute ALL changesets (from the begining of time
+        // to current date) on the database.
 
-    /*
-    for (const changeset of changesets) {
-        const match = changeset.name.match(/^(\d{14})/);
+        console.log("Journal is empty, all changesets are pended.");
 
-        if (!match) {
-            continue;
-        }
+        result = changesets;
+    } else if (changesets.some(changeset => changeset.date < firstExecutedDate)) {
+        // this is a rare case and should not normally happen.
+        // it means that we have old changesets that were not executed on the database.
+        // this is an alarming situation.
 
-        if (!lastExecutedDate || changeset.date > lastExecutedDate) {
-            config.debug2(`\t\tadded changeset ${chalk.yellow(changeset.name)}`);
+        console.warn(chalk.yellow("WARNING: found old changesets before first one."));
+        console.warn(chalk.yellow("\tRestarted changeset execution from the begining of time ..."));
 
-            if (!result.contains(changeset)) {
+        // if we found at least one changeset that is older than the first executed changeset,
+        // we have no choice but to executed ALL changesets on the database, both older ones
+        // and even those that were executed and journaled.
+
+        // we must do this so that the database is placed in a correct state.
+        // (ALL changesets are executed in a CORRECT order).
+
+        // we must perform the execution with the first changeset in timeline, since
+        // we cannot determine how far we shoould go back.
+
+        result = changesets;
+    } else {
+        // we loop through all changesets and add any changeset that is not executed
+        // on the database to the final result (pending changesets).
+        // this seems rational at first glance.
+
+        // however, there is a critical potential bug lurking here.
+        // when we have two or more teams with distinct workflows (each team has their
+        // own dev branches), changeset of a team can damage another team's work.
+        // sadly there is no workaround for such problem that could be avoided
+        // automatically by a tool such as pdcsc.
+        // the only workaround is manual synchronization through face-to-face cooperation
+        // among the teams.
+        // so, it is recommended for teams to manually merge other teams' dev branches into
+        // their dev branch and then try to merge their dev branch to main branch.
+
+        // anyhow, when looping through changeset, if we find a changeset prior to
+        // last executed changeset, we don't care anymore that next changesets are executed
+        // on the database or not. from that point forward, we must executed the changesets
+        // on the database.
+
+        let foundOldChangeset = false;
+
+        for (const changeset of changesets) {
+            const foundExecutedChangeset = executedChangesets.find(cs => changeset.name.equals(cs.name)) != null;
+
+            if (foundOldChangeset) {
+                if (!foundExecutedChangeset) {
+                    config.debug2(`\t\tadded missing changeset ${chalk.yellow(changeset.name)}`);
+                } else {
+                    config.debug2(`\t\tre-executing changeset ${chalk.yellow(changeset.name)}`);
+                }
+
+                result.push(changeset);
+
+                continue;
+            }
+
+            if (!foundExecutedChangeset) {
+                if (changeset.date < lastExecutedDate) {
+                    foundOldChangeset = true;
+
+                    config.debug2(`\t\t${chalk.red("CONFLICT POINT")}`);
+                }
+
+                config.debug2(`\t\tadded missing changeset ${chalk.yellow(changeset.name)}`);
+
                 result.push(changeset);
             }
         }
+
     }
-    */
 
-    result.sort((a, b) => a.date - b.date);
+    if (result.length) {
+        config.debug2("Pending Changesets", result.map(changeset => changeset.name));
 
-    config.debug2("Pending Changesets", result.map(changeset => changeset.name));
-
-    if (result.length === 0) {
-        console.log("No pending changeset found. Database is up-to-date.");
-    } else {
         console.log(`${result.length} changesets found.`);
     }